Integrated farseer.  Created transform component that physics, kinect, and render rely on.  There is currently a hack for shape of the object.  There is a glitch with when someone was moving, then leaves the kinects view.

+++ Source/Kinectitude/Core/Base/IUpdateable.cs
@@ -3,6 +3,6 @@ namespace Kinectitude.Core.Base
 {    
     public interface IUpdateable
     {
-        void OnUpdate(double frameDelta);
+        void OnUpdate(float frameDelta);
     }
 }
